pyproject.toml
[tool.poetry]
name = "pokeonta-bot"
version = "0.1.0"
description = ""
authors = ["Zech Zimmerman <hi@zech.codes>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.4"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
